requirements.txt
django
pyserial
phonenumbers
prettytable
sweetify
pandas
websocket-client
mysql-connector-python
mysqlclient 
websockets
# goto-statement
